remove parameter duplication
